% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SASfromR.R
\name{applySAS}
\alias{applySAS}
\title{Wrapper around SASfromR() to make input data set the first argument.}
\usage{
applySAS(x = NULL, sas_code, ...)
}
\arguments{
\item{x}{R data set, or named list of data sets to be used in SAS.}

\item{sas_code}{A raw string (or vector of raw strings) containing SAS code to be executed.}

\item{...}{Arguments passed on to `SASfromR()`.}
}
\value{
A data set or named list of data sets, corresponding to those produced by SAS and placed in the "out"-library (unless otherwise specified in `outdata`. If no data sets are to be returned, SASfromR will return NULL (silently).
}
\description{
In `SASfromR()`, `sas_code` is the first argument as it's primary purpose
is to execute SAS code (whether you want to attach R data sets or not).
However, in many cases, it is more natural to have the input data set(s) as
a first argument. This wrapper also makes it more easily usable in pipes.
}
\examples{
# Run PROC MEANS on the mtcars data set and return results.
# Note: When a single data set is supplied it will obtain the name "indata" in SAS.
\dontrun{
applySAS(mtcars, r"(
  proc means data=indata;
    class vs;
    var mpg;
    ods output Summary = out.means;
  run;
)")
}

# Example utilizing applySAS() in a pipe
\dontrun{
mtcars |>
  dplyr::filter(am==0) |>
  applySAS(r"(
   proc means data=indata;
     class vs;
     var mpg;
     ods output Summary = out.means;
   run;
 )") |>
 flextable::flextable()
 }
}
