% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{check_sas_log}
\alias{check_sas_log}
\title{Check the SAS log for errors and warnings.}
\usage{
check_sas_log(sas_log)
}
\arguments{
\item{sas_log}{Full path to a stored SAS log.}
}
\value{
A list of errors and.
}
\description{
Check the SAS log for errors and warnings.
}
\examples{
\dontrun{
logfile <- tempfile()
SASfromR("proc contents data=sashelp.cars; run;", log_file = logfile, display_log=FALSE)
display_SAS_log(logfile)}
}
