% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SASfromR.R
\name{checkSASnames}
\alias{checkSASnames}
\title{Check if, and optionally change, names so that they adhere to SAS standards.}
\usage{
checkSASnames(x, type = "columns", repair = TRUE, warn = TRUE, xpt_version = 8)
}
\arguments{
\item{x}{Character vector of names to be checked.}

\item{type}{Character, either "columns" or "dataset", referring to which type of SAS-names are targeted. In xpt version 5, column names may only be 8 characters long, otherwise 32.}

\item{repair}{Logical, whether to repair the names (change them to adhere to SAS standards) or to leave as is.}

\item{warn}{Logical, whether warning should be displayed when names have been repaired.}

\item{xpt_version}{Numeric, which version of xpt (SAS transport files) should be used (version 8 is default and recommended).}
}
\value{
A character vector of names. If repair=FALSE then returns the original names, if repair=TRUE returns the repaired names.
}
\description{
Check if, and optionally change, names so that they adhere to SAS standards.
}
\examples{
# check if the names of an R data set are SAS compliant.
checkSASnames(iris)
# check and repair the names of an R data set.
iris2 <- checkSASnames(iris, repair=TRUE)
}
