% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{display_SAS_log}
\alias{display_SAS_log}
\title{Display a SAS log on the R console.}
\usage{
display_SAS_log(sas_log)
}
\arguments{
\item{sas_log}{Full path to where the SAS log has been stored.}
}
\value{
The SAS log.
}
\description{
Display a SAS log on the R console.
}
\examples{
\dontrun{
logfile <- tempfile()
SASfromR("proc contents data=sashelp.cars; run;", log_file = logfile, display_log=FALSE)
display_SAS_log(logfile)}
}
