% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{display_SAS_output}
\alias{display_SAS_output}
\title{Display the output from SAS procedures}
\usage{
display_SAS_output(sas_output)
}
\arguments{
\item{sas_output}{Full path to where the output has been stored.}
}
\value{
The SAS output.
}
\description{
Display the output from SAS procedures
}
\examples{
\dontrun{
outfile <- tempfile()
SASfromR("proc contents data=sashelp.cars; run;", output_file = outfile, display_output=FALSE)
display_SAS_output(outfile)}
}
