% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/execute_SAS_script.R
\name{execute_SAS_script}
\alias{execute_SAS_script}
\title{Execute SAS code on your machine (requires installation of SAS)}
\usage{
execute_SAS_script(
  sas_script,
  sas_path = NULL,
  sas_output = tempfile(),
  sas_log = tempfile(),
  display_output = TRUE,
  display_log = TRUE
)
}
\arguments{
\item{sas_script}{Full path to a stored SAS script to be executed.}

\item{sas_path}{Full path to sas.exe in your local installation. Only necessary if the directory containing sas.exe is not in your PATH.}

\item{sas_output}{Full path to where the output from SAS procedures should be stored.}

\item{sas_log}{Full path to where the SAS-log should be stored.}

\item{display_output}{Logical, whether to display the output produced by SAS.}

\item{display_log}{Logical, whether to display the log produced by SAS.}
}
\value{
A numeric exit-code.
}
\description{
Execute SAS code on your machine (requires installation of SAS)
}
\examples{
\dontrun{
# create temporary file containing SAS code
sas_script <- tempfile()
readr::write_lines("proc contents data=sashelp.cars; run;", file=sas_script)
# run the stored script using execute_SAS_script()
execute_SAS_script(sas_script)}
}
