% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{export_R_data}
\alias{export_R_data}
\title{Export R data to SAS transport file}
\usage{
export_R_data(indata = NULL, in_path, xpt_version = 8, repair_names = TRUE)
}
\arguments{
\item{indata}{Input data, either a single data.frame or a named list of data.frames.}

\item{in_path}{The full path to the directory where you wish to store it.}

\item{xpt_version}{Version of xpt transfer file to use. Default (and recommended) is version 8.}

\item{repair_names}{Logical, whether to repair/change column names in order to comply with SAS standards (a warning will be issued to tell you which names have been changed and what the new names are).}
}
\value{
A character string that can be used in SAS programs to import the data.
}
\description{
Export R data to SAS transport file
}
\examples{
export_R_data(mtcars,in_path=tempdir())
}
