% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export_R_data.R
\name{fct2fmt}
\alias{fct2fmt}
\title{Create a SAS format data set based on factor variables in an R data.frame.}
\usage{
fct2fmt(df)
}
\arguments{
\item{df}{R data.frame (or objects that inherits from data.frame, e.g. tibble)}
}
\value{
A data.frame which can be exported and used in SAS proc format with the cntlin= option to set formats for factor variables.
}
\description{
Create a SAS format data set based on factor variables in an R data.frame.
}
\examples{
fct2fmt(iris)
}
