% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_SAS_data.R
\name{fmt2fct}
\alias{fmt2fct}
\title{Convert attached formats to proper factor variables}
\usage{
fmt2fct(df, fmts)
}
\arguments{
\item{df}{data.frame (or object that can be coerced into data.frame) with SAS formats attached (as is the case after import via haven).}

\item{fmts}{data.frame of formats (should contain variables FMTNAME, START and LABEL)}
}
\value{
data.frame with factor variables.
}
\description{
Convert attached formats to proper factor variables
}
\examples{
\dontrun{
#not even trying
}
}
