% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{plot_sas_figures}
\alias{plot_sas_figures}
\title{Plot SAS-generated images in plot-pane.}
\usage{
plot_sas_figures(dir)
}
\arguments{
\item{dir}{Path to directory where SAS-generated images are stored.}
}
\description{
Plot SAS-generated images in plot-pane.
}
\examples{
\dontrun{# no examples yet}
}
