% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export_R_data.R
\name{RtoSAS}
\alias{RtoSAS}
\title{Export a single data.frame to R (with optional format statment)}
\usage{
RtoSAS(
  df,
  sas_name,
  libname = NULL,
  directory = tempdir(),
  xpt_version = 8,
  format_statement = NULL,
  ...
)
}
\arguments{
\item{df}{R data.frame to be exported to SAS}

\item{sas_name}{Character. SAS name of data set}

\item{libname}{Character. The libname to be used for the transport data set (must be at most 8 characters).}

\item{directory}{Full file path to where temporary xport data is stored.}

\item{xpt_version}{Which version of xpt to use, 5 or 8 (recommended).}

\item{format_statement}{Optional format statement.}

\item{...}{Not used.}
}
\value{
SAS script to import into SAS
}
\description{
Export a single data.frame to R (with optional format statment)
}
\examples{
RtoSAS(mtcars,"mtcars")
}
