% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SASfromR.R
\name{SASfromR}
\alias{SASfromR}
\title{Execute SAS code from within R.}
\usage{
SASfromR(
  sas_code,
  indata = NULL,
  outdata = NULL,
  xpt_version = 8,
  sas_path = NULL,
  base_dir = tempdir(),
  output_file = NULL,
  log_file = NULL,
  display_log = FALSE,
  display_output = TRUE,
  remove_tempfiles = TRUE,
  repair_names = TRUE
)
}
\arguments{
\item{sas_code}{A raw string, or vector of raw strings, containing the SAS code to be executed.}

\item{indata}{A data set, or a named list of data sets, to be made available to SAS in the "work"-library. If a single data set is provided, it will be accessible under the name "indata". If a named list of datasets is provided, they will be available under their respective names in the named list.}

\item{outdata}{A character vector of the names of data sets you wish SASfromR to return. By default, `outdata=NULL`, all data sets placed in the "out"-library will be returned.}

\item{xpt_version}{Numeric. Which version of SAS transport files (xpt) do you wish to use (version 8 is default and recommended).}

\item{sas_path}{The full path to sas.exe (useful if it is not in your user PATH variable, though ideally you should add it to your PATH). Example: `sas_path=file.path("C:","Program Files","SASHome","SASFoundation","9.4","sas.exe")`.}

\item{base_dir}{The full path to a directory where temporary (if remove_tempfiles=TRUE) files are stored. Defaults to the R sessions temporary directory.}

\item{output_file}{Path to where you wish to store the output (other than generated data sets) from your SAS call. By default the output will be displayed in the R console but not stored persistently. Output to console can be turned off using `display_output=FALSE`.}

\item{log_file}{Path to where you wish to store the log generated by SAS. By default the log will not be stored persistently. To view the log in console, choose `display_log=TRUE`.}

\item{display_log}{Logical, whether to display the SAS log in the R console.}

\item{display_output}{Logical, whether to display the SAS output in the R console.}

\item{remove_tempfiles}{Logical, whether to remove temporary files (e.g. SAS transport files, .sas7bdat-files, generated .sas-files, and log- or output-files unless they have been given more persistent paths through `output_file` and `log_file` arguments). Note: `remove_tempfiles=TRUE` is useful for debugging, but is not intended to be used on a regular basis. Temporary files are automatically removed when the current R-session ends.}

\item{repair_names}{Logical, whether to automatically repair the names of incoming data sets to adhere to SAS standards. If `FALSE` non-compliant names will throw an error.}
}
\value{
A data set or named list of data sets, corresponding to those produced by SAS and placed in the "out"-library (unless otherwise specified in `outdata`. If no data sets are to be returned, SASfromR will return NULL (silently).
}
\description{
A function which allows you to run SAS from within R (through batch-mode),
while using input data sets from R (argument `indata`).
Displays the SAS output in the R console and returns (a subset of, if optional argument `outdata` is used) the data sets produced in SAS and stored in the temporary "out"-library.
}
\details{
SAS code is provided in the `sas_code` argument (preferably as a raw-string).
Input data sets from R will be placed in the SAS library "work" with names taken from the
named list supplied in `indata` (or simply "indata" if a single data set is supplied) and can be referenced in the `sas_code`.
Data sets produced in SAS that you want to output to R should be placed in the SAS library "out" (which is created by the function)
and should thus be written as out."name_of_data_set" in `sas_code`. If only a subset of the data sets in "out" should be returned, this can be specified in the `outdata` argument.
}
\examples{
# Run PROC MEANS on existing SAS data set sashelp.cars, without output
# this will only return the output to console.
\dontrun{
SASfromR(r"(
  proc means data=sashelp.cars;
    class type;
    var mpg_highway;
  run;
)")
}

# If an error occurs (which causes a non-zero exit), usually SASfromR will also output the SAS log.
# If you want to view the log anyway, you can do so with the option display_log = TRUE
\dontrun{
SASfromR(r"(
  proc means data=sashelp.cars;
    class type;
    var mpg_highway;
  run;
)", display_log=TRUE)
}

# To make SASfromR return the results, output to out.means
\dontrun{
SASfromR(r"(
  proc means data=sashelp.cars;
    class type;
    var mpg_highway;
    ods output Summary = out.means;
  run;
)")
}

# Run PROC MEANS on R data set mtcars by using indata argument.
# Note that when a single data set is supplied, it will obtain the name indata
# (or work.indata) in SAS.
\dontrun{
SASfromR(mtcars, r"(
  proc means data=indata;
    class vs;
    var mpg;
    ods output Summary = out.means;
  run;
)")
}

# If you want to chose the name of the input data set in SAS, you
# can do so by wrapping it in a named list.
\dontrun{
SASfromR(list("mtcars"=mtcars), r"(
  proc means data=mtcars;
    class vs;
    var mpg;
    ods output Summary = out.means;
  run;
)")
}
# With named lists, you can make multiple R data sets available to SAS.
# Note that since the R data set iris contains names which are not SAS-compliant,
# these will be repaired and a warning will be issued telling you which names have
# been changed and what the new names are.
\dontrun{
SASfromR(list("mtcars"=mtcars, "iris"=iris), r"(
  proc means data=mtcars;
    class type;
    var mpg_highway;
    ods output Summary = out.mtcars_means;
  run;
  proc means data=iris;
    class Species;
    var Sepal_Length;
    ods output Summary = out.iris_means;
  run;
)")
}
}
