% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export_R_data.R
\name{toSASnames}
\alias{toSASnames}
\title{Convert names to comply with SAS standards}
\usage{
toSASnames(x, type = "variable", warn = TRUE, xpt_version = 8, ...)
}
\arguments{
\item{x}{Character vector of names to be converted.}

\item{type}{Type of SAS name to convert to (variable, dataset or libname).}

\item{warn}{Logical. Whether to warn the user of changes in names.}

\item{xpt_version}{Version of xpt to use, 5 or 8 (default and recommended).}

\item{...}{Not used.}
}
\value{
A character vector of modified names.
}
\description{
Convert names to comply with SAS standards
}
\examples{
toSASnames(c("this.will.change","but_not_this"))
}
